requirements.txt
fastapi
uvicorn
asyncpg
python-dotenv
pydantic
python-multipart
email-validator
passlib[bcrypt]
bcrypt==4.0.1
python-jose[cryptography]
supabase
httpx
pymongo
huggingface_hub 
# AI & NLP stack
torch
transformers
sentence-transformers
keybert
scikit-learn
numpy
nltk
deep-translator
faiss-cpu
sentencepiece
# PDF processing
PyMuPDF
PyPDF2
pdfplumber
